Fix popup message boxes

+++ src/include/music.h
@@ -6,12 +6,12 @@
 #define num_macs  128
 #define maxsndlen 4096
 
-extern long vocposn [num_samps];
-extern int voclen [num_samps];
-extern int vocrate [num_samps];
+extern int32_t vocposn [num_samps];
+extern int16_t voclen [num_samps];
+extern int16_t vocrate [num_samps];
 extern char *soundmac [num_macs];
-extern long textposn [num_text];
-extern int textlen [num_text];
+extern int32_t textposn [num_text];
+extern int16_t textlen [num_text];
 
 extern int *freq;
 extern int *dur;
@@ -19,7 +19,7 @@ extern int *dur;
 extern int vocflag;
 extern int musicflag;
 extern char *textmsg;
-extern int textmsglen;
+extern int16_t textmsglen;
 
 extern unsigned int xmusicrate;
 extern unsigned int xclockrate;
